src/requirements.txt
Django==4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-apscheduler==0.6.2
drf-yasg==1.21.7
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
gunicorn==21.2.0
SQLAlchemy==2.0.31